extract cam and controller object settings
